keepAlive only looks at playing time for rasberry shorter waits in the sc program

+++ keepAlive.sh
@@ -11,21 +11,28 @@ sleep 60 # give it time to start
 
 while true
     do
-
-        if [ -f $playing ] ; then
-            # see if we've been playing too long
-            last_played_a_new_sound=$(( $(date +%s) - $(date +%s -r $playing) ))
-            if [ $last_played_a_new_sound -gt $too_long ] ; then
-                 if [ -f $playing ] ; then # double check
-                    echo "we're stuck playing one thing for too long"
-                    kill $1
-                    #make extra sure
-                    killall jackd
-                    exit 0
+        if [ -f /etc/rpi-issue ]
+            then
+            #only if we're on a pi!
+            if [ -f $playing ] ; then
+                # see if we've been playing too long
+                last_played_a_new_sound=$(( $(date +%s) - $(date +%s -r $playing) ))
+                if [ $last_played_a_new_sound -gt $too_long ] ; then
+                     if [ -f $playing ] ; then # double check
+                        echo "we're stuck playing one thing for too long"
+                        kill $1
+                        #make extra sure
+                        sleep 2
+                        killall scsynth
+                        sleep 0
+                        killall jackd
+                        sleep 1
+                        killall jackd
+                        exit 0
+                    fi
                 fi
             fi
         fi
-
 
         if [ ! -f $alive ]; then
             echo "File not found! - scweets has not checked in and must be hung"
